split reusable form configs into grouped files
Add related objects form config using reusable forms.
